Move types to better places
