+++ packages/evercamps/src/components/frontStore/checkout/cart/items/Items.tsx
@@ -10,60 +10,11 @@ import Quantity from './Quantity';
 import EditParticipantForm from './EditParticipantForm';
 import { useModal } from '@components/common/modal/useModal';
 import { Form } from '@components/common/form/Form';
-
-interface Price {
-  value: number;
-  text: string;
-}
-
-interface CartItemRegistration {
-  cartItemRegistrationId: string;
-  cartItemId: string;
-  firstName: string;
-  lastName: string;
-  extraData?: string;
-  editApi: string;
-  removeApi: string;
-}
-
-interface CartItem {
-  cartItemId: string;
-  thumbnail?: string;
-  qty: number;
-  productName: string;
-  productSku: string;
-  variantOptions?: string;
-  productCustomOptions?: string;
-  productUrl: string;
-  productPrice: Price;
-  productPriceInclTax: Price;
-  finalPrice: Price;
-  finalPriceInclTax: Price;
-  lineTotal: Price;
-  lineTotalInclTax: Price;
-  removeApi: string;
-  updateQtyApi: string;
-  manageRegistrations?: number;
-  registrations: CartItemRegistration[];
-  errors: string[];
-}
-
-interface ParticipantCheckoutField {
-  code: string;
-  label: string;
-  type: 'text' | 'date' | 'select';
-  required: boolean;
-  useForUniqueness: boolean;
-}
-
-interface Setting {
-  priceIncludingTax: boolean;
-  participantCheckoutFields?: string;
-}
+import type { CartItem, CartItemRegistration, CheckoutSetting, ParticipantCheckoutField } from '../../../../../types/checkout';
 
 interface Props {
   items: CartItem[];
-  setting: Setting;
+  setting: CheckoutSetting;
 }
 
 export default function Items({ items, setting: { priceIncludingTax, participantCheckoutFields } }: Props) {
