+++ packages/evercamps/src/modules/catalog/pages/frontStore/productView/ParticipantForm.tsx
@@ -3,14 +3,7 @@ import { Card } from '@components/admin/cms/Card';
 import Button from '@components/common/form/Button';
 import { Field } from '@components/common/form/Field';
 import { _ } from '../../../../../lib/locale/translate/_.js';
-
-interface ParticipantCheckoutField {
-  code: string;
-  label: string;
-  type: 'text' | 'date' | 'select';
-  required: boolean;
-  useForUniqueness: boolean;
-}
+import type { ParticipantCheckoutField } from '../../../../../types/checkout';
 
 interface SavedParticipant {
   participantId: number;
