+++ packages/evercamps/src/types/form.ts
@@ -1,0 +1,2 @@
+export type SelectOption = { value: string | number; text: string };
+export type ValidationRule = string | { rule: string; message?: string };
